c function pointers

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -16,7 +16,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	long int i;
+	unsigned int i;
 
 	if (array == NULL || action == NULL)
 		return;
